Move Die and Dice to their own mods
