+++ src/die.rs
@@ -1,0 +1,115 @@
+use std::ops::Add;
+
+use crate::Rollable;
+
+/// Represents a single die.
+/// Has an initial random value.
+///
+/// # Examples
+/// ## Adding roll values
+/// ```
+/// use one_d_six::Die;
+///
+/// let mut d4 = Die::new(4);
+/// let mut d6 = Die::new(6);
+///
+/// let d4_result: u32 = d4.roll();
+/// let d6_result: u32 = d6.roll();
+/// let result = d4_result + d6_result;
+///
+/// assert!(result >= 2);
+/// assert!(result <= 10);
+/// ```
+///
+/// ## Adding current values
+/// ```
+/// use one_d_six::Die;
+///
+/// let d4 = Die::new(4);
+/// let d6 = Die::new(6);
+///
+/// let d4_result: u32 = d4.current_face();
+/// let d6_result: u32 = d6.current_face();
+/// let result = d4_result + d6_result;
+///
+/// assert!(result >= 2);
+/// assert!(result <= 10);
+/// ```
+///
+/// ## Adding dice directly
+/// ```
+/// use one_d_six::Die;
+///
+/// let d4: Die = Die::new(4);
+/// let d6 = Die::new(6);
+///
+/// let result = d4 + d6;
+///
+/// assert!(result >= 2);
+/// assert!(result <= 10);
+/// ```
+pub struct Die<T: Rollable = u32> {
+    faces: T,
+    current_value: T,
+}
+
+impl<T: Rollable> Die<T> {
+    /// Creates a single die with the specified number of faces.
+    ///
+    /// # Example
+    ///
+    /// ```
+    /// use one_d_six::Die;
+    ///
+    /// let coin: Die = Die::new(2);
+    /// ```
+    pub fn new(faces: T) -> Self {
+        let die = Die {
+            faces,
+            current_value: T::roll(faces),
+        };
+        die
+    }
+    /// Rolls a single die.
+    ///
+    /// # Example
+    ///
+    /// ```
+    /// use one_d_six::Die;
+    ///
+    /// let mut d6: Die = Die::new(6);
+    ///
+    /// assert!(d6.roll() >= 1);
+    /// assert!(d6.current_face() <= 6);
+    /// ```
+    pub fn roll(&mut self) -> T {
+        self.current_value = T::roll(self.faces);
+        self.current_value
+    }
+    /// Gets the current value of the die.
+    ///
+    /// # Example
+    ///
+    /// ```
+    /// use one_d_six::Die;
+    ///
+    /// let d4: Die = Die::new(4);
+    ///
+    /// assert!(d4.current_face() >= 1);
+    /// assert!(d4.current_face() <= 4);
+    /// ```
+    pub fn current_face(&self) -> T {
+        self.current_value
+    }
+}
+
+impl<T: Rollable> Add for Die<T>
+where
+    T: Add,
+{
+    type Output = T::Output;
+
+    fn add(self, other: Self) -> Self::Output {
+        self.current_value + other.current_value
+    }
+}
